move and rename random modules to util dir
